Fix - search nodes by providerID if nodeRef is nil during delete
Fixes an issue that can occur if a machine is deleted before
its nodeRef was set, causing an orphaned node to stay in the
cluster indefinitely with a status of NotReady.

This change modifies isDeleteNodeAllowed to try to find a machine's
associated node if nodeRef is nil by matching the node's providerID
to the machine's. If a match isn't found, the node delete is
skipped, as before.